エントリのメタデータ作成処理をEntryクラスに移動
